Created a default encoder.
The default encoder is intended to provide a default method for
converting class objects to dictionaries, thereby providing support
for a large class of object.

Additionally, this adds support for user-specified custom encodings
using the `__snapshot__` method. Also this add filtering for any
types were a default encoding is not obvious, such as types,
functions, and iterators.

To accomplish this, all encoding information was organized and moved
to `snappiershot/serializers/constants.py`.