Bug Fix -- Partial File Writing
This fixes a bug where if a snapshot file is being written and an
error occurs during serialization, the file is left half written
and therefore unparsable.

The fix writes to a temporary file and moves to the snapshot file
location if no errors occurred, otherwise deleting the temporary
file.